Added support for limited cache size Fixed issues in transaction system Added cache size statistics

+++ src-core/swift/crdt/AddOnlySetCRDT.java
@@ -37,7 +37,7 @@ public class AddOnlySetCRDT<V> extends BaseCRDT<AddOnlySetCRDT<V>> {
     }
     
     @Override
-    public int estimatedSize() {
+    public long estimatedSize() {
         return elements.size();
     }
